requirements.txt
# Core — install transformers from source for latest model support
# pip install git+https://github.com/huggingface/transformers
accelerate
bitsandbytes
torch
pillow

# Qwen-VL image/video preprocessing utilities
qwen-vl-utils[decord]

# Dataset loading
datasets

# Evaluation & visualization
matplotlib
pandas

# PDF handling
PyMuPDF

# Image preprocessing (skew detection / deskew)
opencv-python-headless
numpy

# Gradio UI
gradio

# ───────────────────────────────────────────────────────────
# Colab quick-install (copy into first cell):
#
#   !pip install -q git+https://github.com/huggingface/transformers
#   !pip install -q accelerate bitsandbytes qwen-vl-utils[decord] datasets
#   !pip install -q pillow PyMuPDF opencv-python-headless numpy gradio matplotlib pandas
#
# Llama-3.2-11B-Vision requires HuggingFace login:
#   !huggingface-cli login --token YOUR_TOKEN
# and accepting the license at https://huggingface.co/meta-llama/Llama-3.2-11B-Vision-Instruct
